backend/requirements.txt
asgiref==3.9.1
astroid==3.3.11
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
dill==0.4.0
Django==5.2.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-filter==25.1
django-log-viewer==1.1.8
django-select2==8.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isort==6.0.1
mccabe==0.7.0
openpyxl==3.1.5
packaging==25.0
platformdirs==4.3.8
psycopg2-binary==2.9.10
PyJWT==2.10.1
pylint==3.3.7
pylint-django==2.6.1
pylint-plugin-utils==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tomlkit==0.13.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
